move used visualization functions to the visualizer.py and remove newly created visualization.py
